+++ .github/workflows/uv_publish.yaml
@@ -24,7 +24,7 @@ jobs:
     # Select environment based on manual input, default to 'pypi' on pushed tags
     environment: ${{ inputs.index || 'pypi' }}
     permissions:
-      contents: read
+      contents: write
       id-token: write # Crucial for PyPI Trusted Publishing (OIDC)
 
     steps:
@@ -35,6 +35,39 @@ jobs:
         uses: astral-sh/setup-uv@v7
         with:
           enable-cache: true
+
+      - name: Ensure version tag exists on manual dispatch
+        if: ${{ github.event_name == 'workflow_dispatch' }}
+        shell: bash
+        run: |
+          VERSION="$(python - <<'PY'
+          import tomllib
+          from pathlib import Path
+
+          with Path("pyproject.toml").open("rb") as f:
+              data = tomllib.load(f)
+
+          print(data['project']['version'])
+          PY
+          )"
+
+          TAG="v$VERSION"
+
+          if [[ "$VERSION" == *".dev"* ]]; then
+            echo "Error: Refusing manual publish for development version '$VERSION'."
+            echo "Bump to a stable version before dispatching this workflow."
+            exit 1
+          fi
+
+          echo "Resolved version tag: $TAG"
+
+          if git ls-remote --tags origin "refs/tags/$TAG" | grep -q .; then
+            echo "Tag '$TAG' already exists on origin. Skipping tag creation."
+          else
+            echo "Tag '$TAG' does not exist on origin. Creating and pushing."
+            git tag "$TAG"
+            git push origin "$TAG"
+          fi
 
       - name: Build and Publish to PyPI or TestPyPI
         run: |
