+++ pyproject.toml
@@ -1,13 +1,12 @@
 [project.urls]
-Homepage = "https://github.com/treeolivetech/pkg-tawala"
-Documentation = "https://github.com/treeolivetech/pkg-tawala#readme"
 Repository = "https://github.com/treeolivetech/pkg-tawala"
 Issues = "https://github.com/treeolivetech/pkg-tawala/issues"
-# Changelog = "https://github.com/treeolivetech/pkg-tawala/blob/main/CHANGELOG.md"
+# Homepage = "https://github.com/treeolivetech/pkg-tawala"
+# Documentation = "https://github.com/treeolivetech/pkg-tawala#readme"
 
 [project]
-name = "tawala"
-version = "1.3.1.dev3"
+name = "Tawala"
+version = "1.3.2"
 description = "Build and deploy Django apps with confidence."
 readme = "README.md"
 license = "MIT"
@@ -28,7 +27,7 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 dependencies = [
-    "christianwhocodes==1.5.2",
+    "christianwhocodes==1.5.4",
     "django==6.0.3",
     "django-browser-reload==1.21.0",
     "django-compressor==4.6.0",
@@ -51,8 +50,8 @@ psycopg = ["psycopg[binary,pool]==3.3.3"]
 dev = [
     "djlint>=1.36.4",
     "psycopg[binary,pool]==3.3.3",
-    "ruff>=0.15.6",
-    "ty>=0.0.23",
+    "ruff>=0.15.7",
+    "ty>=0.0.24",
     "vercel==0.5.2",
 ]
 
@@ -61,7 +60,7 @@ tawala = "tawala.management.cli:main"
 twl = "tawala.management.cli:main"
 
 [build-system]
-requires = ["uv_build>=0.10.11,<0.11.0"]
+requires = ["uv_build>=0.10.12,<0.11.0"]
 build-backend = "uv_build"
 
 [tool.ruff]
