+++ src/tawala/contrib/__init__.py
@@ -1,1 +1,1 @@
-"""Apps module."""
+"""Contrib apps module."""
