+++ src/tawala/contrib/admin/apps.py
@@ -2,7 +2,7 @@
 
 from django.apps import AppConfig
 
-from ...constants import Package
+from ... import Package
 
 
 class AdminConfig(AppConfig):
