+++ src/tawala/management/settings/_13_sass.py
@@ -1,6 +1,6 @@
 """Sass Configuration."""
 
-from ...constants import InstalledApps
+from ... import InstalledApps
 from ._07_installed_apps import INSTALLED_APPS
 
 if InstalledApps.SASS_PROCESSOR in INSTALLED_APPS:
